+++ database/migrations/2017_11_25_005034_create_entry_option_table.php
@@ -17,8 +17,8 @@ class CreateEntryOptionTable extends Migration
     public function up()
     {
         Schema::create('entry_option', function (Blueprint $table) {
-            $table->integer('entry_id')->unsigned()->index();
-            $table->integer('option_id')->unsigned()->index();
+            $table->bigInteger('entry_id')->unsigned()->index();
+            $table->bigInteger('option_id')->unsigned()->index();
 
             $table->foreign('entry_id')->references('id')->on('entries')->onDelete('cascade');
             $table->foreign('option_id')->references('id')->on('options')->onDelete('cascade');
