+++ database/migrations/2018_03_24_154539_create_manual_votes_table.php
@@ -18,8 +18,8 @@ class CreateManualVotesTable extends Migration
     {
         Schema::create('manual_votes', function (Blueprint $table) {
             $table->id();
-            $table->integer('competition_id')->unsigned()->index();
-            $table->integer('entry_id')->unsigned()->index();
+            $table->bigInteger('competition_id')->unsigned()->index();
+            $table->bigInteger('entry_id')->unsigned()->index();
             $table->string('points');
             $table->string('ip_address');
             $table->timestamps();
