+++ database/migrations/2018_03_28_155604_add_vote_category_id_to_votes.php
@@ -17,7 +17,7 @@ class AddVoteCategoryIdToVotes extends Migration
     public function up()
     {
         Schema::table('votes', function (Blueprint $table) {
-            $table->integer('vote_category_id')->after('competition_id')->unsigned()->index();
+            $table->bigInteger('vote_category_id')->after('competition_id')->unsigned()->index();
             $table->foreign('vote_category_id')->references('id')->on('vote_categories')->onDelete('cascade');
 
             $table->boolean('special_vote')->after('entry_id');
