+++ database/migrations/2018_03_31_172238_add_final_file_to_entries.php
@@ -17,7 +17,7 @@ class AddFinalFileToEntries extends Migration
     public function up()
     {
         Schema::table('entries', function (Blueprint $table) {
-            $table->integer('final_file_media_id')->after('competition_id')->unsigned()->nullable()->index();
+            $table->bigInteger('final_file_media_id')->after('competition_id')->unsigned()->nullable()->index();
             $table->foreign('final_file_media_id')->references('id')->on('media')->onDelete('set null');
         });
     }
