+++ database/migrations/2018_12_01_151704_create_component_entry_uploads_table.php
@@ -17,7 +17,7 @@ class CreateComponentEntryUploadsTable extends Migration
     {
         Schema::create('component_entry_uploads', function (Blueprint $table) {
             $table->id();
-            $table->integer('entries_page_id')->unsigned()->nullable()->index();
+            $table->bigInteger('entries_page_id')->unsigned()->nullable()->index();
             $table->timestamps();
         });
     }
